power: Introduce S4
Enlighten the lower level portions of powerd to the existence of
hibernate, also known as S4 or suspend-to-disk. With this change, powerd
is able to orchestrate a hibernate operation, though there are currently
no upper-level facilities yet to ask it to do so.

A new default_pref, disable_hibernate, is set initially to 1 to ensure
none of this code is accidentally lit up on existing machines.
Eventually once things are fully baked, we'll want to remember to flip
disable_hibernate's default to 0 so that each new platform doesn't have
to override it.

Note, the original resource for shutdown_from_suspend.png is here:
https://docs.google.com/drawings/d/1pD7lnyABKkx53hBgrMl7hdgzkcWvRi_Pzr5b0_5leL8/edit

BUG=b:192360150
TEST=Try suspend-to-disk on dedede and volteer

Change-Id: I3691d61db945654cdae4b08056e4f7574098374c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3017216
Tested-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Reviewed-by: C Shapiro <shapiroc@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>