patchpanel: eliminate spurious log noise
When patchpanel starts for the first time, it is expected that the
iptables rule
  "-A OUTPUT -m owner ! --uid-owner 1000 -j skip_apply_vpn_mark"
will not have been created yet, and cannot "successfully" be deleted.

Since this operation is part of initializing the iptables state to a
known starting point, this is not a failure per say and the error can be
ignored like other such operations trying to delete custom jump rules in
Datapath::ResetIptables().

BUG=None
TEST=unit tests. Flashed DUT, rebooted, checked that spurious error log
is gone from /var/log/net.log.

Change-Id: I2ebbeb432c91a695cf0692a6e6f97fa259616813
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3016446
Tested-by: Hugo Benichi <hugobenichi@google.com>
Auto-Submit: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Jason Jeremy Iman <jasongustaman@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>