metrics: Add per-device keys file for structured metrics
We're adding support for events in the chromium instance of structured
metrics to be keyed by a device-wide ID. This means keys must be saved
somewhere outside the cryptohome for device-wide IDs.

This CL creates the /var/lib/metrics/structured/chromium/ directory
for this purpose, which the keys file will be saved into in
crrev.com/c/2981431. This creates a directory, rather than touching a
file, so that chromium can continue to use ImportantFileWriter to save
keys. This writes to a temporary file which is then moved into place,
so needs rwx permissions on the directory itself. This has 755
permissions for consistency with the parent directory, though we only
expect access by chronos so could be 700.

I've also updated all tmpfiles.d entries here to use d= rather than d,
for consistency with the other metrics/tmpfiles.d/ file.

Note the StatefulFiles tast test doesn't need to be updated, as the
exception for /var/lib/metrics is still in place. It makes sense for us
to leave it there until the lacros migration is finished, as that may
also affect these files.

BUG=chromium:1148168
TEST=CQ, and required directory created on DUT.

Change-Id: Iabbe3087771b5b5b3d83bd7b62c7c8c7b60dc8b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3016454
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Tested-by: Tony Yeoman <tby@chromium.org>
Commit-Queue: Tony Yeoman <tby@chromium.org>