lorgnette: Add enum values for new backends
canon_lide70 was previously reported as "other", and epsonds is a
newly-added backend.  While we're touching these, also switch to the
newer recommended SendEnumToUMA function.  Because we were previously
incorrectly using an inclusive max value instead of exclusive, this will
cause the previous max enum bucket to be reported with two separate
names.  This is a tiny fraction of the overall counts, so allowing the
double-reporting is better than renaming all our metrics.

BUG=b:188600541
TEST=emerge-zork lorgnette; unit tests

Change-Id: I89837fd64c71e9077a617af8b5a729bd9a0a4359
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3022453
Reviewed-by: Mason Wilde <masonwilde@google.com>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>