cryptohome: Change AddKeyset behavior to not reuse a VaultKeyset.
This CL changes the way how VaultKeysets are used during AddKeyset
operation. Previously we loaded the old keyset, reencrypted it and
resaved it as new one. This left some old data in the keyset. This CL
creates a fresh VK and the copies data that is needed.

BUG=chromium:1220890
TEST=FEATURES=test emerge-${BOARD} cryptohome
TEST=tast tests - crytohome*, hwsec*
TEST=manual test, create pin, log out, log back in

Change-Id: I6473864c2d69ca26074fc095ce7512feb0e52939
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3025244
Tested-by: Hardik Goyal <hardikgoyal@chromium.org>
Auto-Submit: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>