init: Mount enc stateful and rollback pstore file
Mount and unmount encrypted stateful during clobber to retrieve rollback
file and an update engine file that clobber already tried to fetch
before but it never did.

Move the rollback file to pstore because it needs to be preserved
across powerwash but must not be stored on unencrypted stateful.

BUG=chromium:1212958
TEST=Together with other CLs in this chain:
     1) Deploy on device and run a full rollback with powerwash.
     2) FEATURES=test emerge-volteer oobe_config
     3) FEATURES=test emerge-volteer chromeos-init

Change-Id: I6aa23aebe78bcc28e6bde8c2ec88cbf8aa403f7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3027120
Reviewed-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-by: Amr Aboelkher <amraboelkher@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Commit-Queue: Miriam Polzer <mpolzer@google.com>
Tested-by: Miriam Polzer <mpolzer@google.com>