missive: Prevent UploadJob from checking.
The reason is that job may be destructed by SharedQueue, and weak
pointer in UploadJob is deleted on a different seq task runner than
the expected one (that belongs to scheduler). To prevent that, I made
it possible to specify the runner when creating SharedQueue - so now
both Scheduler and SharedQueue use the same runner.

There is currently no test that repros the issue, I've added TODO to
prepare it (it must fail if SharedQueue is created with new task
runner).

BUG=b:193577465
TEST=FEATURES="test" emerge-$BOARD missive

Change-Id: I1ca31dc4ea81c1723f4725b0017ff730f68c6343
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3027701
Reviewed-by: Kyle Williams <kdgwill@chromium.org>
Tested-by: Leonid Baraz <lbaraz@chromium.org>
Commit-Queue: Leonid Baraz <lbaraz@chromium.org>