power: Document/codify RequestSuspend parameters
The XML bindings for RequestSuspend were missing wakeup_timeout, an
optional uint32 popped out in HandlRequestSuspendMethod(). Document
that. For wakeup_timeout, passing a value of 0 is already equivalent to
not passing one.

Also allow callers to pass in a reserved -1 value for
external_wakeup_count to indicate not to match on wakeup count (in other
words the same thing as not passing it).

This will be useful when adding a third optional parameter specifying
the suspend type (hibernate or resume).

BUG=b:192360150
TEST=emerge-volteer power_manager

Change-Id: I6a0d5cc8e824f9f40279129f588da09ba496e587
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3042619
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Tested-by: Evan Green <evgreen@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>