chromeos-config: add support for mtk wifi power tables
In order to support dynamic SAR functionality on Mediatek devices, add a config
to store the dynamic wifi power tables in chromeos-config.  The specs of these
tables are not compatible with the rtw or ath10k driver tables, so add a
separate table definition, and require that devices can only define one type of
table using the "oneOf" keyword.

BUG=b:188752295, b:147757591
TEST=./run_tests.sh
TEST=manual: Add non-tablet-mode-power-table-mtk to
     overlay-asurada/chromeos-base/chromeos-config-bsp-asurada/files/model.yaml,
     `FEATURES=test sudo -E emerge chromeos-base/chromeos-config-host,
     `emerge-asurada chromeos-config-bsp-asurada  chromeos-config` succeeds.

Change-Id: I1e47954a741c8823e2ac59239a8225316bcd4835
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3032429
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Tested-by: Joshua Emele <jemele@chromium.org>
Commit-Queue: Joshua Emele <jemele@chromium.org>