shill: Remove OnNewWiphy band error log
Before chromium:3028791, we would expect an error if the band message
didn't contain any frequency attributes. Now that it's possible for the
message to split, it's possible for the split message with band information to
not have frequency attributes. We drop this error in this change.

BUG=b:192279407
TEST=tast run ${DUT} wifi.RandomMACAddress on dedede, hatch as well as
unit tests

Change-Id: Ib26370dd49fafabe9182338c5683aa4d6df6e0b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3034598
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Billy Zhao <billyzhao@chromium.org>
Tested-by: Billy Zhao <billyzhao@chromium.org>