shill: dbus-client: Simplify proxy lifetime
This patch eliminates the unnecessary explicit initialization step
performs it instead in the constructor. Previously the manager proxy
would be destroyed if the client disconnected from shill (e.g. on
restart) but is unnecessary and leads to potential use-after-free
crashes (e.g. dns-proxy trying to set a manager property). This patch
ensures there are no null pointers for object proxies, which was the
underlying cause of the reported crashes.

Users of this client should ensure that shill is indeed ready by hooking
the appropriate callback and tracking when/if shill is reset or lost.

BUG=b:193090160
BUG=b:193712306
TEST=units
TEST=revert crrev.com/c/3032326, apply dns-proxy patch, cannot repro
crash

Change-Id: Ic6d3b49f13f43696795909db4d9aba2b65a9c391
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3037764
Tested-by: Garrick Evans <garrick@chromium.org>
Commit-Queue: Garrick Evans <garrick@chromium.org>
Reviewed-by: Jason Jeremy Iman <jasongustaman@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>