dns-proxy: Fix shill client usage
In short, the shill client does not need to be cleared and reset when/if
the service itself is detected to be lost. Rather, we just need to keep
track of it to avoid using it in these cases. So this patch ensures that
the OnAvailableCallback is hooked if shill is lost and initializes the
client just once on setup.

BUG=b:193712306
BUG=b:193090160
TEST=units
TEST=revert seccomp patch, with new shill client patch, cannot repro
crash

Change-Id: I260681157952905ab26aeccb384fa8718d551e9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3037765
Tested-by: Garrick Evans <garrick@chromium.org>
Reviewed-by: Jason Jeremy Iman <jasongustaman@chromium.org>
Commit-Queue: Garrick Evans <garrick@chromium.org>