debugd: Add features to feature list only if device supports them
We need this so that Chrome can set experiments as "Disabled" but should
be doing so only if the device supports the experiment.

In short, the enabled and disabled groups are activated only for devices
that support an experiment, not for all devices. This makes the samples
collected more relevant and comparable.

BUG=b:179777419
TEST=Run dbus-send on command line to get list.

Change-Id: Id978c17b73506316a5a816e5d4a01827e9a1a6eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3039198
Reviewed-by: Joel Fernandes <joelaf@google.com>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Joel Fernandes <joelaf@google.com>
Tested-by: Joel Fernandes <joelaf@google.com>