iioservice: disable EC polling when sensor is disabled
For cros-ec sensors, when we set ODR frequency to 0, also disable EC
polling. It prevents the EC from sending interrupt to the AP with an
empty FIFO. It also prevents older firmware to report samples with a
different timestamp that they use to.

BUG=b:193093133
TEST=Check when we shutdown (and stop ui), that iioservice sends a
command to set EC period to 0. [on cyan with a modified firmware with
more tracing].

Change-Id: I0677357bc9bc7f974343267990ffcf9e4f49550a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3039559
Reviewed-by: Cheng-Hao Yang <chenghaoyang@chromium.org>
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>
Tested-by: Gwendal Grignou <gwendal@chromium.org>