shill: cellular: Avoid sending repetitive metrics
When using auto connect on a cellular device, and the device keeps
failing to connect because of the same reason, we should avoid sending
the same metrics over and over to avoid having a incorrect
representation of those metrics in UMA.

BUG=b:194180446
TEST=Successfully connected with Lazor and then injected an error in MM
to make all successive connections fail. The device only logged one
metric for the first failure, and nothing after the following failures.

Change-Id: I353df99b39672720d080a88157c39247b236513b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3043347
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>