vm_tools: maitred: Make the LXD 4.0 transition persistent
LXD upgrades cannot be reverted without deleting everything and
starting over. To ensure that we never try to go backwards, have
maitred write a magic file to the stateful disk that records this
transition has occurred when it is requested via a feature flag, and
use the existence of the file (rather then the transient feature flag)
to control which version of LXD is started.

This file is created in the root of the stateful disk which is owned
by root, so we shouldn't have issues with anyone trying to delete it.

This should make it safe to plumb this up to a higher level
control (such as finch) without fear of breaking someone if they move
out of the experiment group.

BUG=chromium:1097610
TEST=Enabled feature flag, booted VM, disabled, booted again, checked
that LXD 4.0 was used afterwards.

Change-Id: I6bf1bf5a3f7ee270e5a529997ccc5b391eb91172
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3041305
Commit-Queue: Fergus Dall <sidereal@google.com>
Tested-by: Fergus Dall <sidereal@google.com>
Reviewed-by: David Munro <davidmunro@google.com>