shill: cellular: Log when loading values from storage
When some settings or flags(i.e. Attach APN) are enabled in Guest Mode,
the user or tester sometimes assumes that the feature will survive a
reboot. Troubleshooting these issues are hard because the settings are
different from what the tester or user is describing in the bug. Log the
cellular settings loaded on statart so we know what to expect.

BUG=b:186253508
TEST=None

Change-Id: Ic400d4280c51faf7b63f7fa314af47114a4d60f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3046319
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Auto-Submit: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>