federated: switched to an example iterator model.
The current example reading model requires clients to keep track of the
state of the database / storage manager and manually sync their reads.

This CL refactors example reading into an iterator model, which is
cleaner and easier to use.

This CL doesn't compile on its own; it is part of a chain of interrelated
CLs necessary to introduce this functionality.

BUG=chromium:1020051
TEST=portage test stage on final CL in chain

Change-Id: If5be737e50427fdb3feca720c8ed1e2bf517768f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3045830
Tested-by: Michael Martis <martis@chromium.org>
Commit-Queue: Michael Martis <martis@chromium.org>
Reviewed-by: Xinglong Luan <alanlxl@chromium.org>