federated: example iterators for storage manager.
This change is the next in the chain of CLs switching the daemon over to
an iterator model of example reading. An iterator model is clearer for
clients and simpler to implement.

This CL doesn't compile on its own; it is part of a chain of interrelated
CLs necessary to introduce this functionality.

BUG=chromium:1020051
TEST=portage test stage on final CL in chain

Change-Id: Ie4bfb4bedf53abb8c2d59d801d1630384cfed3f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3045831
Tested-by: Michael Martis <martis@chromium.org>
Commit-Queue: Michael Martis <martis@chromium.org>
Reviewed-by: Xinglong Luan <alanlxl@chromium.org>