missive: Force events upload upon queue init
Do this if a queue is not empty - which is important if the device was
offline and is coming online after reboot. It is important for
non-periodic queues (Immediate and Manual), since they might not get
another chance to upload any time soon. But it won't harm periodic
queues either, so we do it for all queues.
Also fix the test to not access storage_ when not needed.

BUG=b:194134378
TEST=FEATURES="test" emerge-$BOARD missive

Change-Id: I3e57425b2072a4eb8119d0c3a25f681a7b403d3e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3047718
Auto-Submit: Leonid Baraz <lbaraz@chromium.org>
Tested-by: Leonid Baraz <lbaraz@chromium.org>
Reviewed-by: Kyle Williams <kdgwill@chromium.org>
Commit-Queue: Leonid Baraz <lbaraz@chromium.org>