webserver: Support libmicrohttpd upgrade 0.9.42 to 0.9.72
Need to update int to MHD_Result starting 0.9.71 to build without
failures.
In 0.9.71, the enum MHD_Result was introduced, so any new builds would
fail for many functions that previously used the int.
In addition, this CL gets rid of a check for a fd that we don't use.

BUG=b:194342674
TEST=built clean image using current 0.9.42 and updated 0.9.72 versions
TEST=https://jetstream-autotest.corp.google.com/afe/#tab_id=view_job&object_id=2579131

Change-Id: I54d870b68049489724575c55b3ccc0a2634588d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3059532
Reviewed-by: Aileen Cheng <chengaileen@google.com>
Reviewed-by: John Riordan <jrio@google.com>
Reviewed-by: John Riordan <jrio@chromium.org>
Tested-by: Aileen Cheng <chengaileen@google.com>
Commit-Queue: Aileen Cheng <chengaileen@google.com>