Enable packet_capture in verified-boot mode
Currently packet_capture command is only available in developer mode.
Enable it in verified mode in Crosh. Only device-based (L3) capture will
be available in verified-boot mode and frequency-based capture options
will be available in developer mode. Add checks to the DBus method
implementation in debugd to make sure the frequency-based capture is
only available in dev-mode. Update the debugd DBus method documentation
to explain the options that are only available for dev-mode.
The change is made to implement the feature request in the attached bug.

BUG=b:181525070
TEST=Open Crosh and run "packet_capture" command in verified mode with
--device option.

Change-Id: Ib5e7f199d08cf83f33e239f8ba91841849be8a2d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3058892
Tested-by: Irem Uğuz <iremuguz@google.com>
Commit-Queue: Irem Uğuz <iremuguz@google.com>
Reviewed-by: Allen Webb <allenwebb@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>