vm_tools:  /dev/kmsg logs without throttling.
Sets printk.devkmsg=on for termina VM.
I tried setting this in the StartVMRequest proto, but
this (correctly) means the VM is untrusted. Short of
allowing a blessed set of kernel params to be passed,
setting this in concierge for termina is simplest.

Testing locally on nami and zork show no significant increase
in log output when using crostini. On the plus side,
we see all startup messages from maitred.

BUG=chromium:1233821
TEST=On DUT, crostini starts, all maitred messages in crosvm log

Change-Id: I5706a45b8567dd393f5f9fef9c17e7d2300b757b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3056700
Tested-by: Nicholas Verne <nverne@chromium.org>
Auto-Submit: Nicholas Verne <nverne@chromium.org>
Commit-Queue: Fergus Dall <sidereal@google.com>
Reviewed-by: Fergus Dall <sidereal@google.com>