crosh: Add privacy note to the output of packet_capture
Add a note to the user to inform them that the output file of the
packet_capture command can contain personal information and not to share
the file with untrusted parties. The change was suggested by privacy
review for the feature in the attached bug.

BUG=b:181525070
TEST=Open Crosh and run "packet_capture" command. There must be an
informing note at the end of command output.

Change-Id: I70e9c075ebb4110990ab75fc4116f0b446cb9e20
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3062963
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Irem Uğuz <iremuguz@google.com>
Commit-Queue: Irem Uğuz <iremuguz@google.com>