shill: Support domain_suffix_match
domain_suffix_match is used by the wpa_supplicant to be matched against
the dNSName element of the SubjectAltName of the authentication server
certificate.

shill receives the domain_suffix_match from Chrome as a list of strings
and serializes it as a semicolon separated string of entries to send it
to the wpa_supplicant.

DD: go/secure-eap-wifi-match

BUG=b:177651194,chromium:1203612
TEST=shill unit tests
TEST=tast test wifi.SimpleConnect.* (CL:3063299)

Change-Id: I4277b22ae1bb14bc5ee9addd51fd3fb12d21a0a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3062970
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Andreea-Elena Costinas <acostinas@google.com>
Commit-Queue: Andreea-Elena Costinas <acostinas@google.com>