lorgnette: support B&W JPEG scans
Add logic to the JPEG reader to support black and white scans. When the
JPEG reader sees data with a bit depth of 1, it will expand each bit in
the input data to a byte before writing the expanded data to libjpeg.

BUG=b:194224377
TEST=FEATURES=test emerge-nami lorgnette, plus manual testing with the
scan app on a DUT.

Change-Id: If681154ae7992b47da9ab022dc89ffd22ec14e82
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3064580
Commit-Queue: Paul Moy <pmoy@chromium.org>
Tested-by: Paul Moy <pmoy@chromium.org>
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>