common-mk: Make platform2_test handle empty or absent /etc/passwd
platform2_test.py sets up /etc/passwd in the sysroot with an entry for
the calling user, but it currently assumes that /etc/passwd is present
and has at least one line. This was previously okay because by the
time unit tests are run many users should have been added by various
packages.

However with crrev.com/c/3062147 it can also be run by meson builds to
execute unit tests. When meson is setting up the build directory it
attempts to build and run a test binary to check that the configured
exe_wrapper actually works, and this can happen very early in the
build process. In particular, the package pixman uses meson and
actually has *no* dependencies in the sysroot, so this can happen
before /etc/passwd actually exists or has any entries.

This CL removes both assumptions, creating /etc/passwd if it is not
present and allowing it to be empty.

BUG=chromium:1227738
TEST=emerge-tatl pixman with empty /etc/passwd and no /etc/passwd

Change-Id: If27ec4808480fe17ba96a5adf93ed6c40f259239
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3063863
Tested-by: Fergus Dall <sidereal@google.com>
Commit-Queue: Fergus Dall <sidereal@google.com>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>