power: add BatteryStatePoll signal
In order to be compatible with fwupd, the BatteryStatePoll
signal has been added to the power_manager. It is redundant
with PowerSupplyPoll in terms of what information it gives
about the system, but it gives the information in a tuple
with three elements instead of as a protobuf, which is how
PowerSupplyPoll returns its information. This way, fwupd does
not need to have a dependency on protobuf, and it is not
affected by potential changes to orderings or added entries
in the method call it calls.

The three properties returned by BatteryStatePoll are the
external power type, the battery state, and the display
battery percentage. External power type is used so that
"battery discharging" and "AC power connection" can be
viewed as separate events. The battery state is sent as
an enum matching the upower enum implementation. This
way, the method call is not dependent on system_api
changes not directly related to the method call. The
display percentage was chosen over the actual percentage
since they are very close in number, but it is easier to
test manually with the display percentage and it would be
helpful in the case that fwupd develops a UI.

BUG=b:195450074
TEST=from ssh: dbus-monitor --system interface='org.chromium.PowerManager'
  output of test:
   uint32 1
   uint32 4
   double 100
from chroot: emerge-volteer power_manager ; cros_workon_make power_manager --test --board=volteer

Change-Id: Ia7d2a0ef97e2517a1be73552d1fd0da6f59cee70
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3072559
Tested-by: Twain Byrnes <binarynewts@google.com>
Reviewed-by: Daniel Campello <campello@chromium.org>
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Commit-Queue: Twain Byrnes <binarynewts@google.com>