system_api: balloon_policy in StartArcVmRequest
To allow Chrome (and a Finch experiment) to control the new
virtio_balloon sizing policy, add an optional parameter to
StartArcVmRequest. If present, it will enable the new policy configured
with the provided parameters. If not present, the existing balance
available policy will be used.

BUG=b:194632077
TEST=tast run dut multivm.Login.arc

Change-Id: I56f11699abf3c9b3b4efe6270334e13fa4a5c3b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3077111
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-by: Hikaru Nishida <hikalium@chromium.org>
Tested-by: Charles William Dick <cwd@google.com>
Commit-Queue: Charles William Dick <cwd@google.com>