vm_tools: concierge: bump startup timeout to 120s
Attempting to start a Bruschetta instance using OVMF as the BIOS on some
classes of hardware (an eve board in my case) sometimes ends up timing
out because the startup time is very close to the current 60s limit.

Previously this was doubled to 30s to 60s when BIOS support was added to
concierge in
https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2772916
but startup times have slowed to the point where we can't reliably beat
the 60s timeout.

This doubles the timeout to make VM startup for these types of VMs
possible.

BUG=b:194954660
TEST=built and deployed to eve test device and confirmed that the
increased timeout caused successful boot

Change-Id: I7813f8f6b332c9f0322e803e092ad8951424a384
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3086191
Commit-Queue: Chirantan Ekbote <chirantan@chromium.org>
Tested-by: Mike Gerow <gerow@google.com>
Auto-Submit: Mike Gerow <gerow@google.com>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>