dns-proxy: Fix dns-proxy crash on timeout
dnsproxyd might crash if a callback is triggered after
its destructor is called. This is because the curl multi
handle is destroyed on its class destructor.

This happens because the callbacks are called in an unsafe
manner. Add a weak pointer to ensure that the callbacks
are canceled when that class is destroyed.

This patch also adds a curl multi guard in the timeout
function to be safe.

BUG=b:194862268
TEST=call destructor while a query is running;
     observe that dnsproxyd does not crash;
     without this patch dnsproxyd crash;

Change-Id: If00ccd7e04319d36d9c14b7ade2dfffd548b4f97
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3089275
Tested-by: Jason Jeremy Iman <jasongustaman@chromium.org>
Reviewed-by: Garrick Evans <garrick@chromium.org>
Commit-Queue: Jason Jeremy Iman <jasongustaman@chromium.org>