installer: fix postinst when no firmware update is possible
FirmwareUpdate is supposed to return an integer exit code (with 0
meaning success), but when chromeos-firmwareupdate doesn't exist it
incorrectly returns a boolean true, which is converted to 1 which is
treated as a failure.

This appears to have been caused by a refactoring bug in
crrev.com/c/49338, where the return type was changed from bool to int.

This causes a problem on the reven board, where no firmware updater
exists. Signed images have /root/.force_update_firmware, so this code
path is hit but then fails when chromeos-firmwareupdate is not found.

BUG=b:194885246
TEST=Build a reven image. Run it and "touch /root/.force_update_firmware", then verify installation succeeds.

Change-Id: Ia2d9cc9158e81bbef2cbbc2d4cf93d980dbdcef4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3088479
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>