diagnostics: Provide udev foundation
In this CL, we create the net link for udev events before entering the
sandbox. The purpose is to avoid dropping the network namespace.

Besides, we need to enable those syscall in seccomp policy for libudev
library.

For udev foundation, we create a udev events monitor, it watches all
udev events. We can use |subsystem| and |action| to check if this is our
interested events.

BUG=none
TEST=Test with other CLs

Change-Id: I1c12229dc8e1ba0423c6d67a3f82dd6bf41b7f26
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3085698
Reviewed-by: Chung-Sheng Wu <chungsheng@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>
Tested-by: Kerker Yang <kerker@chromium.org>
Commit-Queue: Kerker Yang <kerker@chromium.org>