shill: vpn: Introduce VPNConnection base class
This patch introduces a VPNConnection class, which will be the base
class for IPsecConnection and L2TPConnection. A VPNConnection manages
a VPN connection: its owner can call Connect()/Disconnect() and register
callbacks on it, and its derived class should implement OnConnect()/
OnDisconnect() and trigger Notify*() to invoke the callbacks on events.

The functionality of VPNConnection is a subset of that of VPNDriver.
Their main differences are:
- VPNConnection does not involve PropertyStore and Profile, while
  VPNDriver does (so we cannot use VPNDriver for IPsecConnection and
  L2TPConnection);
- VPNConnection is supposed to be created before connecting, and
  destroyed after disconnected, while the lifetime of VPNDriver is same
  as Service.

BUG=b:165170125
TEST=unittest

Change-Id: I1155c2025fc41ec6dfac6f567bdb617e901263c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3096733
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>