arc/vmtools: seccomp: allow newfstatat syscall
Starting with glibc 2.33 newfstatat is used more on 64bit
architectures which support it, in our case amd64 and arm64.

Previously we have added it with a script doing bulk addition
based on pre-existence of fxstatat{64} which covered the vast
majority of uses, but these two are undetected outliers due to
not using fxstatat{64} so we add them now.

BUG=b:187795855
TEST=Local builds; CQ.

Change-Id: I20a8294ae83082546b7c038cb7ebe17e6e4a889e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3097753
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-by: Edman Anjos <edman@chromium.org>
Tested-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: Manoj Gupta <manojgupta@chromium.org>