vm_tools: seneschal: Switch to using "expect daemon"
Currently the double invocation to minijail leads to the second
minijail process being considered the "main" process of seneschal,
which means we have to keep this process around just to forward
signals from upstart and wait on seneschal exiting.

In order to make it safe to exit the second minijail process, we need
it to not be running as init in a PID namespace (if init exits, all
processes in that PID namespace get hit by SIGKILL). This is okay, the
outer PID namespace is redundant as minijail always sets up a PID
namespace when setting up a user namespace.

This reveals a quirk with /proc, however. Currently we have five
different options set that touch this: -r to remount /proc as
read-only, -k to mount as read-write, -p to set up a PID
namespace (which implies -r), and --profile=minimalistic-mountns,
which effectively does -rd, ro bind / to /, ro bind /dev/log to
/dev/log, ro bind /proc to /proc (yes, this one option sets /proc to
read-only in *two different ways*), pivot-root to /mnt/empty, and
finally -I to run as init, which causes /proc to be left untouched.

Currently the last one in that list takes precedence (I think?) but
removing it leads to a ro /proc, which breaks user namespacing (we
need to be able to write to /proc/pid/uid_map while setting up the
inner jail). I have dealt with this by replacing the --profile option
with the options it sets up, and then removing all the ro /proc
options (since we actually *don't* want them).

-U implies -pv (pid namespace, mount namespace), but I add them there
to be explicit. Normally it would also imply remounting /proc as ro,
but this is skipped because we run as init. Since we do want it, I do
this explicitly using -k in the inner jail.

Also put seneschal in a net namespace, it doesn't use networking.

BUG=chromium:1240118
TEST=Manually tested

Cq-Depend: 3097231
Change-Id: I279a39ffa0921d4c30f05d4cee9a16274efe0187
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3096894
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Tested-by: Fergus Dall <sidereal@google.com>
Commit-Queue: Fergus Dall <sidereal@google.com>