arc: Optionally mount dedicated /cache partition
A dedicated /cache partition is required to be mounted in the first
stage of the init process for the adb remount / sync to work on Android
R+ running on ARCVM. This mount is only required when /cache has a
dedicated partition (i.e. when CACHE_PARTITION is true in the config).

The primary use of the /cache partition is to use it as the upper layer
of overlayfs for the adb remount and sync.

In order for this feature to work, the /cache partition is required to
be a dedicated partition, since it needs to be mounted before the other
partitions before hand.
It is implicitly required in adb remount code, where it checks
if "/cache" is present in the mount entries (ReadDefaultFstab reads
current stab entries, and GetEntryForMountPoint checks it one by one).

The file system could be anything that supports overlayfs. Using
virtiofs is also an option, but it requires Kernel v5.10, as well as
couple of modifications to crosvm that needs further consideration (we
are planning to change it to this approach in near future). Squashfs is
read only, so it is not supported as mount point for the upper layer.

For this reason, we need a different disk formatted to file system that
supports overlayfs, in this case, we chose ext4.

BUG=b:182953041
TEST=chroot$ FEATURES=test emerge-$BOARD arc-setup

Change-Id: I23fe8410b713e6a1025de33c3a38481d85d88627
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3099425
Reviewed-by: Daichi Hirono <hirono@chromium.org>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Tested-by: Keita Suzuki <suzukikeita@google.com>
Auto-Submit: Keita Suzuki <suzukikeita@google.com>
Commit-Queue: Yusuke Sato <yusukes@chromium.org>