tpm_manager: Change a fatal error to error code
When `TpmConnection::GetContext()` returns an empty context, it does not
return an error code. In fuzzing, the destructor of `ScopedTssType` will
report an fatal error of "Leaking Trousers memory due to null context".

Make it return an error code to avoid this problem.

BUG=b:195262420
TEST=tast run hwsec.* on relm

Change-Id: Ifdacc32cff136dda6384e83ea80a0345dc448586
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3099412
Tested-by: Domen Su <domen@google.com>
Commit-Queue: John L Chen <zuan@chromium.org>
Reviewed-by: John L Chen <zuan@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>