crash: gzip devcoredump files
These files can sometimes be megabytes in size. It's best to gzip them
so that we can upload them to crash and not waste tons of bandwidth for
users. This also avoids problems where crash/ doesn't accept files that
are larger than 1.2MB. For a qcom video devcoredump this reduces the
size from 5MB to 490kB, which is a good amount of savings. Presumably
there are a bunch of zeroes in the devcoredumps for BSS, etc. so this is
usually a good idea.

BUG=b:172987695
TEST=cros_workon_make --test --board=amd64-generic crash-reporter

Cq-Depend: chromium:3105372
Change-Id: I976d76f4baa25d0bc9454fbe85f69c790f69360c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3101776
Tested-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>