shill: vpn: Add skeleton code for IPsecConnection::OnConnect()
This patch adds the skeleton code for the connecting procedure in
IPsecConnection class, which describes the relationship among tasks.
Implementation for each task will be added in the follow-up patches.

Also changes the implementation for VPNConnection::NotifyFailure(),
since in a derived class, OnDisconnect() will always be called with
NotifyFailure().

BUG=b:165170125
TEST=unit_tests

Change-Id: I54b9b658e45cd81d066e9b3922949a583d559dda
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3103049
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>